proy_mathmatic/requirements.txt
asgiref==3.8.1
Django==5.0.14
django-mssql-backend==2.8.1
mssql-django==1.5
pyodbc==5.2.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary

